backend_servicer/requirements/requirements.txt
python-decouple
loguru
grpcio
grpcio-tools
scipy
numpy